wsl: usernet: Don't always bind to port 2222
podman-machine starts gvproxy without specifying -ssh-port.
This means it will always bind port 2222 (see
https://github.com/containers/gvisor-tap-vsock/issues/402), effectively
bypassing the dynamic ssh port allocation.

I don't think wsl needs this port to be opened/bound on the host, but
this commit starts gvproxy with -ssh-port set to the dynamic SSH port to
avoid port conflicts with crc.
When gvproxy allows to disable this port forwarding entirely, we can
revise this.

This should fix #20327

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>