cirrus: never run functional tests on rhel branches
As discussed at the cabal October 8, 2024 we have no need for these
tests on RHEL branches. The work to maintain them is higher than it is
worth it. We also do not test RHEL but rather some outdated frozen
fedora image build from the time we created the branch.

Therefore we gain little value from them especially as all the interal
Red Hat QE is testing it anyways again on the proper RHEL builds.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>