cirrus: remove functional tests
As discussed at the cabal October 8, 2024 we have no need for these
tests on RHEL branches. The work to maintain them is higher than it is
worth it. We also do not test RHEL but rather some outdated frozen
fedora image build from the time we created the branch.

Therefore we gain little value from them especially as all the internal
Red Hat QE is testing it anyways again on the proper RHEL builds.

So simply delete all the stuff we no longer need:
 - alt builds, no point in windows/macos testing and other arches
 - all the functional tests
 - the build success task (not needed as there is nothing after it
   anymore)
 - the swagger task, we do not use the swagger from the rhel branches

Fixes: https://issues.redhat.com/browse/RUN-2315

Signed-off-by: Paul Holzinger <pholzing@redhat.com>