[v5.2-rhel] libpod: exposed port use old loop syntax
An additional tweak from @Luap99 in #24333
regarding the looping in libpod/container_inspect.go.

The range over int syntax was only added in go 1.22, this branch is
currently build with go 1.21 in RHEL so we need to convert it back to
the old syntax.

And add the missing "fmt" import in the test file.

Signed-off-by: tomsweeneyredhat <tsweeney@redhat.com>