Exclude symlink from pre-commit end-of-file-fixer
Git on Windows clones the symlink `test/python/requirements.txt`
as a regular file. That's not a big deal except that `pre-commit`,
executed by `winmake.ps1 lint`, complains because there is no new
line at end of the file.

To fix `winmake.ps1 lint` we exclude the symlink in `pre-commit`
hook `end-of-file-fixer` configuration.

Signed-off-by: Mario Loriedo <mario.loriedo@gmail.com>