contrib/systemd: use multi-user.target instead of default.target
The systemd documentation[1] says the following about using
`default.target` as `WantedBy=` in system service files:

  For typical unit files please set "WantedBy=" to a regular target
  (like multi-user.target or graphical.target), instead of default.target,
  since such a service will also be run on special boots like on system
  update, emergency boot ...

The situation for user service files is different.
There default.target is actually the right one to use.

The mentioned "system update" special boots refer to
`systemd.offline-updates`[2] a mechanism that enables doing a special
minimalistic boot to run e.g. migration scripts after installing an
update (using a package manager or an image based updater like e.g. RAUC)
and before the first boot of the updated system.

To prevent conflicts between normal services on the system and migration
scripts, normal services should not be started for these special boots.

For this to work normal services may not use `WantedBy=default.target`,
because according to the documentation[2] the `system-update.target`
_becomes_ the `default.target` for system-update boots:

  3) Very early in the new boot systemd-system-update-generator(8) checks
     whether /system-update or /etc/system-update exists.
     If so, it (temporarily and for this boot only) redirects
     (i.e. symlinks) default.target to system-update.target,
     a special target that pulls in the base system
     (i.e. sysinit.target, so that all file systems are mounted but little
     else) and the system update units.

Use `WantedBy=multi-user.target` target instead of `default.target` to
enable the use of `system-update.target`.

[1]: https://www.freedesktop.org/software/systemd/man/latest/systemd.special.html#default.target
[2]: https://www.freedesktop.org/software/systemd/man/latest/systemd.offline-updates.html#

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>