contrib/systemd/user: replace symlinks with copies of the files
Symlinks in git repositories tend to be confusing - a change in one file
results in a change in a seemingly unrelated location without causing
a diff.

In addition to that the best practices between user service files and
system service files differ when in comes to `WantedBy=`.

Create copies of the service files to allow them to diverge in the future.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>