DO NOT MERGE! Parallelize podman build tests
DO NOT MERGE! Seriously, like don't even fantasize about it
until https://github.com/containers/buildah/issues/5674 is
fixed and the fix is vendored into podman.

I am filing this because if/when that happens, this PR will
give you a nice CI-runtime boost. I spent a good chunk of
time identifying the tests which can / can't be parallelized
in this module. Hope it helps some day.

Signed-off-by: Ed Santiago <santiago@redhat.com>